+++ public/functions-enqueue.php
@@ -1,10 +1,17 @@
 <?php
 
-function arve_enqueue_styles() {
+function arve_register_styles() {
+
+  arve_register_asset( array(
+    'automin' => true,
+    'handle'  => ARVE_SLUG,
+    'src'     => plugin_dir_url( __FILE__ ) . 'arve.css'
+  ) );
+}
+
+function arve_print_maxwidth_style() {
 
   $options = arve_get_options();
-
-  wp_enqueue_style( ARVE_SLUG, plugin_dir_url( __FILE__ ) . 'arve-public.css', array(), ARVE_VERSION, 'all' );
 
   if ( (int) $options["video_maxwidth"] > 0 ) {
     wp_add_inline_style( ARVE_SLUG, sprintf( '.arve-wrapper{max-width:%dpx;}', $options['video_maxwidth'] ) );
@@ -13,11 +20,10 @@ function arve_enqueue_styles() {
 
 function arve_register_scripts() {
 
-  wp_register_script(
-    ARVE_SLUG,
-    plugin_dir_url( __FILE__ ) . 'arve-public.js',
-    array( 'jquery' ),
-    ARVE_VERSION,
-    true
-  );
+  arve_register_asset( array(
+    'automin' => true,
+    'handle'  => ARVE_SLUG,
+    'src'     => plugin_dir_url( __FILE__ ) . 'arve.js',
+    'deps'    => array( 'jquery' )
+  ) );
 }