+++ public/functions-validation.php
@@ -13,6 +13,23 @@ function arve_validate_url( $url ) {
   return false;
 }
 
+function arve_validate_aspect_ratio( $aspect_ratio ) {
+
+	if ( empty( $aspect_ratio ) ) {
+		return $aspect_ratio;
+	}
+
+	$a = explode( ':', $aspect_ratio );
+
+	if ( ! empty( $a[0] ) && is_numeric( $a[0] ) && ! empty( $a[1] ) && is_numeric( $a[1] ) ) {
+		return $aspect_ratio;
+	}
+
+  return new WP_Error( 'Aspect ratio',
+    sprintf( __( 'Aspect ratio <code>%s</code> is not valid', ARVE_SLUG ), $aspect_ratio )
+  );
+}
+
 function arve_bool_to_shortcode_string( $val ) {
 
   if ( false === $val ) {
@@ -43,7 +60,7 @@ function arve_validate_bool( $val, $name ) {
       break;
     default:
       $val = new WP_Error( $name,
-        sprintf( __( '%s <code>%s</code> not valid', 'advanced-responsive-video-embedder' ), $name, $val )
+        sprintf( __( '%s <code>%s</code> not valid', ARVE_SLUG ), $name, $val )
       );
       break;
   }